Move modals files
