+++ src/pages/EditorPage/VsCodeBanner.js
