+++ src/pages/EditorPage/Welcome/index.js
